requirements.txt
telethon
cryptg
pillow
hachoir
requests
aiohttp
aiofiles
python-dotenv
